+++ python/sdssdb/peewee/sdss5db/catalogdb.py
@@ -241,6 +241,13 @@ class Gaia_DR2(CatalogdbModel):
         table_name = 'gaia_dr2_source'
 
 
+class Gaia_DR2_g_lt_18(CatalogdbModel):
+    source_id = BigIntegerField(primary_key=True)
+
+    class Meta:
+        table_name = "gaia_dr2_source_g_lt_18"
+
+
 class Gaia_EDR3(CatalogdbModel):
 
     source_id = BigIntegerField(primary_key=True)
@@ -255,6 +262,13 @@ class Gaia_DR3(CatalogdbModel):
 
     class Meta:
         table_name = 'gaia_dr3_source'
+
+
+class Gaia_DR3_g_lt_16(CatalogdbModel):
+    source_id = BigIntegerField(primary_key=True)
+
+    class Meta:
+        table_name = "gaia_dr3_source_g_lt_16"
 
 
 class Gaia_DR2_Neighbourhood(CatalogdbModel):
@@ -2328,7 +2342,7 @@ class ToO_Target(CatalogdbModel):
 
     class Meta:
         table_name = "too_target"
-        reflect = False
+        use_reflection = False
 
 
 class ToO_Metadata(CatalogdbModel):
@@ -2359,7 +2373,7 @@ class ToO_Metadata(CatalogdbModel):
 
     class Meta:
         table_name = "too_metadata"
-        reflect = False
+        use_reflection = False
 
 
 _Gaia_DR2_TwoMass_Deferred.set_model(Gaia_DR2_TwoMass_Best_Neighbour)
@@ -2386,7 +2400,7 @@ class CatalogFromSDSS_DR19p_Speclite(CatalogdbModel):
         # in "catalog_to_X" tables
         table_name = "catalog_from_sdss_dr19p_speclite"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToAllStar_DR17_synspec_rev1(CatalogdbModel):
@@ -2405,7 +2419,7 @@ class CatalogToAllStar_DR17_synspec_rev1(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_allstar_dr17_synspec_rev1"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToAllWise(CatalogdbModel):
@@ -2424,7 +2438,7 @@ class CatalogToAllWise(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_allwise"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToBHM_CSC(CatalogdbModel):
@@ -2443,7 +2457,7 @@ class CatalogToBHM_CSC(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_bhm_csc"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToBHM_RM_v0(CatalogdbModel):
@@ -2462,7 +2476,7 @@ class CatalogToBHM_RM_v0(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_bhm_rm_v0"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToBHM_RM_v0_2(CatalogdbModel):
@@ -2481,7 +2495,7 @@ class CatalogToBHM_RM_v0_2(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_bhm_rm_v0"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToBHM_eFEDS_Veto(CatalogdbModel):
@@ -2500,7 +2514,7 @@ class CatalogToBHM_eFEDS_Veto(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_bhm_efeds_veto"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToCatWISE(CatalogdbModel):
@@ -2519,7 +2533,7 @@ class CatalogToCatWISE(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_catwise"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToCatWISE2020(CatalogdbModel):
@@ -2538,7 +2552,7 @@ class CatalogToCatWISE2020(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_catwise2020"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToGLIMPSE(CatalogdbModel):
@@ -2557,7 +2571,7 @@ class CatalogToGLIMPSE(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_glimpse"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToGLIMPSE360(CatalogdbModel):
@@ -2576,7 +2590,7 @@ class CatalogToGLIMPSE360(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_glimpse360"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToGUVCat(CatalogdbModel):
@@ -2595,7 +2609,7 @@ class CatalogToGUVCat(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_guvcat"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToGaiaQSO(CatalogdbModel):
@@ -2614,7 +2628,7 @@ class CatalogToGaiaQSO(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_gaia_qso"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToGaia_DR2(CatalogdbModel):
@@ -2633,7 +2647,7 @@ class CatalogToGaia_DR2(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_gaia_dr2_source"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToGaia_DR2_WD_SDSS(CatalogdbModel):
@@ -2652,7 +2666,7 @@ class CatalogToGaia_DR2_WD_SDSS(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_gaia_dr2_wd_sdss"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToGaia_DR3(CatalogdbModel):
@@ -2671,7 +2685,7 @@ class CatalogToGaia_DR3(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_gaia_dr3_source"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToGaia_unWISE_AGN(CatalogdbModel):
@@ -2690,7 +2704,7 @@ class CatalogToGaia_unWISE_AGN(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_gaia_unwise_agn"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToKeplerInput_DR10(CatalogdbModel):
@@ -2709,7 +2723,7 @@ class CatalogToKeplerInput_DR10(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_kepler_input_10"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToLegacy_Survey_DR10(CatalogdbModel):
@@ -2728,7 +2742,7 @@ class CatalogToLegacy_Survey_DR10(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_legacy_survey_dr10"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToLegacy_Survey_DR10a(CatalogdbModel):
@@ -2747,7 +2761,7 @@ class CatalogToLegacy_Survey_DR10a(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_legacy_survey_dr10a"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToLegacy_Survey_DR8(CatalogdbModel):
@@ -2766,7 +2780,7 @@ class CatalogToLegacy_Survey_DR8(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_legacy_survey_dr8"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToMILLIQUAS_7_7(CatalogdbModel):
@@ -2785,7 +2799,7 @@ class CatalogToMILLIQUAS_7_7(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_milliquas_7_7"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToPS1_g18(CatalogdbModel):
@@ -2804,7 +2818,7 @@ class CatalogToPS1_g18(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_ps1_g18"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToPanstarrs1(CatalogdbModel):
@@ -2823,7 +2837,7 @@ class CatalogToPanstarrs1(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_panstarrs1"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSDSS_DR13_PhotoObj(CatalogdbModel):
@@ -2842,7 +2856,7 @@ class CatalogToSDSS_DR13_PhotoObj(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_sdss_dr13_photoobj"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSDSS_DR13_PhotoObj_Primary(CatalogdbModel):
@@ -2865,7 +2879,7 @@ class CatalogToSDSS_DR13_PhotoObj_Primary(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_sdss_dr13_photoobj_primary"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSDSS_DR16_APOGEE_Star(CatalogdbModel):
@@ -2884,7 +2898,7 @@ class CatalogToSDSS_DR16_APOGEE_Star(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_sdss_dr16_apogeestar"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSDSS_DR16_SpecObj(CatalogdbModel):
@@ -2903,7 +2917,7 @@ class CatalogToSDSS_DR16_SpecObj(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_sdss_dr16_specobj"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSDSS_DR19p_Speclite(CatalogdbModel):
@@ -2922,7 +2936,7 @@ class CatalogToSDSS_DR19p_Speclite(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_sdss_dr19p_speclite"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSkies_v1(CatalogdbModel):
@@ -2941,7 +2955,7 @@ class CatalogToSkies_v1(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_skies_v1"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSkies_v2(CatalogdbModel):
@@ -2960,7 +2974,7 @@ class CatalogToSkies_v2(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_skies_v2"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSkyMapper_DR1_1(CatalogdbModel):
@@ -2979,7 +2993,7 @@ class CatalogToSkyMapper_DR1_1(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_skymapper_dr1_1"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSkyMapper_DR2(CatalogdbModel):
@@ -2998,7 +3012,7 @@ class CatalogToSkyMapper_DR2(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_skymapper_dr2"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSuperCosmos(CatalogdbModel):
@@ -3017,7 +3031,7 @@ class CatalogToSuperCosmos(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_supercosmos"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToTIC_v8(CatalogdbModel):
@@ -3036,7 +3050,7 @@ class CatalogToTIC_v8(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_tic_v8"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToTIC_v8_Extended(CatalogdbModel):
@@ -3055,7 +3069,7 @@ class CatalogToTIC_v8_Extended(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_tic_v8_extended"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToTwoMassPSC(CatalogdbModel):
@@ -3074,7 +3088,7 @@ class CatalogToTwoMassPSC(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_twomass_psc"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToTwoqz_sixqz(CatalogdbModel):
@@ -3093,7 +3107,7 @@ class CatalogToTwoqz_sixqz(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_twoqz_sixqz"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToTycho2(CatalogdbModel):
@@ -3112,7 +3126,7 @@ class CatalogToTycho2(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_tycho2"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToUVOT_SSC_1(CatalogdbModel):
@@ -3131,7 +3145,7 @@ class CatalogToUVOT_SSC_1(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_uvotssc1"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToXMM_OM_SUSS_4_1(CatalogdbModel):
@@ -3150,7 +3164,7 @@ class CatalogToXMM_OM_SUSS_4_1(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_xmm_om_suss_4_1"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToXMM_OM_SUSS_5_0(CatalogdbModel):
@@ -3169,7 +3183,7 @@ class CatalogToXMM_OM_SUSS_5_0(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_xmm_om_suss_5_0"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogTounWISE(CatalogdbModel):
@@ -3188,7 +3202,7 @@ class CatalogTounWISE(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_unwise"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToMarvels_dr11_star(CatalogdbModel):
@@ -3207,7 +3221,7 @@ class CatalogToMarvels_dr11_star(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_marvels_dr11_star"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToMarvels_dr12_star(CatalogdbModel):
@@ -3226,7 +3240,7 @@ class CatalogToMarvels_dr12_star(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_marvels_dr12_star"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToSDSS_DR17_SpecObj(CatalogdbModel):
@@ -3245,7 +3259,7 @@ class CatalogToSDSS_DR17_SpecObj(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_sdss_dr17_specobj"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class CatalogToToO_Target(CatalogdbModel):
@@ -3264,7 +3278,7 @@ class CatalogToToO_Target(CatalogdbModel):
     class Meta:
         table_name = "catalog_to_too_target"
         primary_key = False
-        reflect = False
+        use_reflection = False
 
 
 class SDSS_ID_To_Catalog(CatalogdbModel):
@@ -3272,6 +3286,7 @@ class SDSS_ID_To_Catalog(CatalogdbModel):
     sdss_id = BigIntegerField()
     catalogid = BigIntegerField()
     version_id = IntegerField()
+    lead = TextField()
     allstar_dr17_synspec_rev1__apstar_id = TextField()
     allwise__cntr = BigIntegerField()
     bhm_rm_v0__pk = BigIntegerField()
@@ -3284,9 +3299,14 @@ class SDSS_ID_To_Catalog(CatalogdbModel):
     guvcat__objid = BigIntegerField()
     legacy_survey_dr10__ls_id = BigIntegerField()
     legacy_survey_dr8__ls_id = BigIntegerField()
+    mangatarget__mangaid = TextField()
+    marvels_dr11_star__starname = TextField()
+    marvels_dr12_star__pk = BigIntegerField()
+    mastar_goodstars__mangaid = TextField()
     panstarrs1__catid_objid = BigIntegerField()
     ps1_g18__objid = BigIntegerField()
     sdss_dr13_photoobj__objid = BigIntegerField()
+    sdss_dr17_specobj__specobjid = TextField()
     skymapper_dr1_1__object_id = BigIntegerField()
     skymapper_dr2__object_id = BigIntegerField()
     supercosmos__objid = BigIntegerField()
@@ -3343,6 +3363,26 @@ class SDSS_ID_To_Catalog(CatalogdbModel):
         Legacy_Survey_DR8,
         column_name='legacy_survey_dr8__ls_id',
         backref='+')
+    mangatarget = ForeignKeyField(
+        Mangatarget,
+        column_name='mangatarget__mangaid',
+        backref='+'
+    )
+    marvels_dr11_star = ForeignKeyField(
+        Marvels_dr11_star,
+        column_name='marvels_dr11_star__starname',
+        backref='+'
+    )
+    marvels_dr12_star = ForeignKeyField(
+        Marvels_dr12_star,
+        column_name='marvels_dr12_star__pk',
+        backref='+'
+    )
+    mastar_goodstars = ForeignKeyField(
+        Mastar_goodstars,
+        column_name='mastar_goodstars__mangaid',
+        backref='+'
+    )
     panstarrs1 = ForeignKeyField(
         Panstarrs1,
         column_name='panstarrs1__catid_objid',
@@ -3354,6 +3394,10 @@ class SDSS_ID_To_Catalog(CatalogdbModel):
     sdss_dr13_photoobj = ForeignKeyField(
         SDSS_DR13_PhotoObj,
         column_name='sdss_dr13_photoobj__objid',
+        backref='+')
+    sdss_dr17_specobj = ForeignKeyField(
+        SDSS_DR17_SpecObj,
+        column_name='sdss_dr17_specobj__specobjid',
         backref='+')
     skymapper_dr1_1 = ForeignKeyField(
         SkyMapper_DR1_1,
@@ -3387,4 +3431,4 @@ class SDSS_ID_To_Catalog(CatalogdbModel):
     class Meta:
         table_name = "sdss_id_to_catalog"
         primary_key = False
-        reflect = False
+        use_reflection = False