+++ schema/lvmdb/lvmopsdb.sql
@@ -214,16 +214,27 @@ CREATE TABLE lvmopsdb.guider_coadd (
     exposure_no INTEGER
 );
 
-CREATE TABLE lvmopsdb.overhead (
+CREATE TABLE lvmopsdb.ln2_fill (
     pk SERIAL PRIMARY KEY NOT NULL,
-    observer_id BIGINT,
-    tile_id INTEGER,
-    stage TEXT,
-    start_time DOUBLE PRECISION,
-    end_time DOUBLE PRECISION,
-    duration REAL
+    action TEXT,
+    start_time TIMESTAMPTZ,
+    end_time TIMESTAMPTZ,
+    purge_start TIMESTAMPTZ,
+    purge_complete TIMESTAMPTZ,
+    fill_start TIMESTAMPTZ,
+    fill_complete TIMESTAMPTZ,
+    fail_time TIMESTAMPTZ,
+    abort_time TIMESTAMPTZ,
+    failed BOOLEAN,
+    aborted BOOLEAN,
+    error TEXT,
+    log_file TEXT,
+    json_file TEXT,
+    configuration JSONB,
+    log_data JSONB,
+    plot_paths JSONB,
+    valve_times JSONB
 );
-
 
 -- constraints
 
@@ -359,15 +370,23 @@ CREATE INDEX CONCURRENTLY standard_qc3_index
     ON lvmopsdb.standard
     (q3c_ang2ipix(ra, dec));
 
+CREATE INDEX ON lvmopsdb.ln2_fill (action);
+CREATE INDEX ON lvmopsdb.ln2_fill (start_time);
+CREATE INDEX ON lvmopsdb.ln2_fill (end_time);
+CREATE INDEX ON lvmopsdb.ln2_fill (purge_start);
+CREATE INDEX ON lvmopsdb.ln2_fill (purge_complete);
+CREATE INDEX ON lvmopsdb.ln2_fill (fill_start);
+CREATE INDEX ON lvmopsdb.ln2_fill (fill_complete);
+CREATE INDEX ON lvmopsdb.ln2_fill (fail_time);
+CREATE INDEX ON lvmopsdb.ln2_fill (abort_time);
+CREATE INDEX ON lvmopsdb.ln2_fill (failed);
+CREATE INDEX ON lvmopsdb.ln2_fill (aborted);
+
 CLUSTER standard_qc3_index ON lvmopsdb.standard;
 
 CREATE INDEX CONCURRENTLY ON lvmopsdb.guider_coadd (exposure_no);
 CREATE INDEX CONCURRENTLY ON lvmopsdb.guider_frame (exposure_no);
 CREATE INDEX CONCURRENTLY ON lvmopsdb.agcam_frame (exposure_no);
-
-CREATE INDEX CONCURRENTLY ON lvmopsdb.overhead (observer_id);
-CREATE INDEX CONCURRENTLY ON lvmopsdb.overhead (tile_id);
-CREATE INDEX CONCURRENTLY ON lvmopsdb.overhead (stage);
 
 grant usage on schema lvmopsdb to sdss_user;
 
@@ -377,7 +396,8 @@ lvmopsdb.weather,
 lvmopsdb.observation_to_standard,
 lvmopsdb.observation_to_sky,
 lvmopsdb.exposure,
-lvmopsdb.completion_status to sdss_user;
+lvmopsdb.completion_status,
+lvmopsdb.ln2_fill to sdss_user;
 
 grant select on all tables in schema lvmopsdb to sdss_user;
 grant usage on all sequences in schema lvmopsdb to sdss_user;