+++ CHANGELOG.rst
@@ -5,6 +5,16 @@ Changelog
 
 This document records the main changes to the ``sdssdb`` code.
 
+* Add table and models for ``lvmopsdb.ln2_fill``.
+
+* :release:`0.13.0 <2024-09-03>`
+* :feature:`264` Add metadata for the ``sdss_id_to_catalog`` table.
+* :feature:`266` Support PEP 621.
+* :feature:`268` Allows to use `psycopy3 <https://www.psycopg.org/psycopg3/docs/>`__ as a driver for the Peewee database connection.
+* Add missing columns to the ``SDSS_ID_To_Catalog`` models (the columns were being completed via reflection).
+* Change ``reflection=False`` to ``use_reflection=False`` in many ``catalogdb`` models. The previous setting was incorrect and was causing the models to be reflected.
+* Add ``gortdb`` schema and models to ``lvmdb``. Moved ``lvmopsdb.overhead`` to ``gortdb``.
+
 * :release:`0.12.4 <2024-08-07>`
 * Fix ``catalogid`` and ``version_id`` fields in ``SDSS_ID_To_Catalog`` model.